ratings MDL-21657 Improving capabilities for ratings.
Rather than performing complicated upgrade procedures from the old (broken) logic in three activity modules, we are adding new capabilities to those three modules to make them all consistent, and using them in a callback from the Ratings API.  This is the same as the Comments API.

The downside is that two sets of capabilities need to be enabled for ratings to work.  Hopefully this can be mitigated somewhat when build some sort of shortcuts on the module settings forms to set capabilities.  In the meantime, the effect is reduced somewhat by having all the rating capabilities enabled by default for student and teacher roles.  This means control is in the hands of whatever the modules have set, giving very close backward compatibility.

This is not the perfect solution, but neither were the other two.