MDL-69541 mod_forum: Post a copy to all groups in quick discussion
When sitting behind the 'Advanced' link in discussion creation, the 'Post a copy to all groups' option is not very much visible
for a discussion creator. There were many proposals within the community to bring this option to the quick discussion creation
form.

The current improvement brings the 'Post a copy to all groups' option back to the quick discussion creation form.

A behat tests was added to cover the feature's behavior.